agents/slides/requirements.txt
anthropic
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
